requirements.txt
Flask==3.0.2
Flask-WTF==1.2.1
WTForms==3.1.2
Pillow==9.3.0